# Labels: Reduce stringlabels memory usage for common labels