--- a/tsdb/db.go
+++ b/tsdb/db.go
@@ -224,6 +224,9 @@ type Options struct {
 	// PostingsDecoderFactory allows users to customize postings decoders based on BlockMeta.
 	// By default, DefaultPostingsDecoderFactory will be used to create raw posting decoder.
 	PostingsDecoderFactory PostingsDecoderFactory
+
+	// PreInitFunc is a function that will be called before the HEAD is initialized.
+	PreInitFunc PreInitFunc
 }
 
 type NewCompactorFunc func(ctx context.Context, r prometheus.Registerer, l *slog.Logger, ranges []int64, pool chunkenc.Pool, opts *Options) (Compactor, error)
@@ -233,6 +236,8 @@ type BlocksToDeleteFunc func(blocks []*Block) map[ulid.ULID]struct{}
 type BlockQuerierFunc func(b BlockReader, mint, maxt int64) (storage.Querier, error)
 
 type BlockChunkQuerierFunc func(b BlockReader, mint, maxt int64) (storage.ChunkQuerier, error)
+
+type PreInitFunc func(*DB)
 
 // DB handles reads and writes of time series falling into
 // a hashed partition of a seriedb.
@@ -1009,6 +1014,10 @@ func open(dir string, l *slog.Logger, r prometheus.Registerer, opts *Options, rn
 	inOrderMaxTime, ok := db.inOrderBlocksMaxTime()
 	if ok {
 		minValidTime = inOrderMaxTime
+	}
+
+	if db.opts.PreInitFunc != nil {
+		db.opts.PreInitFunc(db)
 	}
 
 	if initErr := db.head.Init(minValidTime); initErr != nil {