@bboreham 2025-02-11
Regarding the hard-coded strings, I don't want this to descend into bike-shedding, but I do think as Kubernetes users we have quite a different profile from you.
From the TSDB Status page on a couple of Prometheus at work I got:

### Label Names
instance 
`__name__ `
job 
pod 
name 
cluster
image 
namespace
uid
status_code
method
container
address
container_id
name
cidr
pod_ip
node_id

### Label Values:
false
200
GET
kube-system/cadvisor
kube-system/kubelet
kube-state-metrics/kube-state-metrics
node-exporter/node-exporter
node-exporter
kube-system

### Metric names
(excluding ones from home-grown software)
kafka_writer_write_wait_seconds_bucket
kafka_writer_write_time_seconds_bucket
storage_operation_duration_seconds_bucket
kube_pod_status_phase
kube_pod_status_ready
container_spec_cpu_period
container_spec_memory_limit_bytes
container_memory_working_set_bytes
container_memory_usage_bytes
container_memory_max_usage_bytes
container_cpu_usage_seconds_total
kube_pod_container_resource_requests
kube_pod_container_status_restarts_total