@bboreham 2025-02-11
> read the last TSDB block on startup and build a mapping from that

Isn't that quite an intensive operation, involving reading the whole index?
Another thought would be to have the WAL checkpoint process emit the table we need.  But maybe it comes out the same.