@bboreham 2025-02-12
Benchmarks: there are a number of slow-downs, including a couple at +50% and one at +1943%, which deserve attention.  Maybe we can code-golf them back a bit.  I'll fire up PromBench so we get an idea whether this is microbenchmark exaggeration or not.
If it's a general slow-down I think we'd have to put this change behind a build tag.

(next time you can skip the regex ones - they don't touch Labels)