@bboreham 2025-02-12
Just looked at the "Top 10 label names with high memory usage" section in 
http://prombench.prometheus.io/15988/prometheus-pr/tsdb-status and http://prombench.prometheus.io/15988/prometheus-release/tsdb-status

We see an improvement on the hard-coded strings.  The overall number is still large because the size of the value is included.

Quoting here so the info is available after prombench is canceled:
### release

Name | Bytes
-- | --
`__name__` | 163682433
nodeName | 162238010
instance | 84573660
job | 74249041


### pr
Name | Bytes
-- | --
nodeName | 162238010
`__name__` | 137938564
job | 67725287
instance | 61740521