@GiedriusS 2025-02-18
> > Some downstream users might want to transfer this data over the wire
> 
> How would you do that? I can see how you would do it inside the Labels package, but then you don't need an export.

Through the Bytes() method https://github.com/prometheus/prometheus/blob/main/model/labels/labels_stringlabels.go#L62. Without exporting this new mapping, it's unclear what the bytes mean.