@bboreham 2025-02-18
`Bytes` is opaque; it is not supposed to be a serialization/deserialization mechanism. 

I’ve done a whole symbol-table implementation, under `-tags dedupelabels`. I think it’s better to improve that (or write a 4th implementation) rather than slowly turn `stringlabels` into that.