@bboreham 2025-02-19
I was responding to @GiedriusS, not on the changes in this PR.

Building a list from the last TSDB block is acceptable, though I might still do it as an add-on PR.

However, we can use PromBench to check the impact.  I think I can restart the process so it will see a block.