@bboreham 2025-02-20
The reason to leave tags as-is was to avoid inconvenience for downstream users such as Thanos.
Probably best to open a separate issue to propose changing the default.