@bboreham 2025-02-24
I took some time to look at the performance of this PR over the weekend, and tried to speed it up.

However I think this kind of thing is a blocker:

> Builder-20                                          423.9n ± ∞ ¹   4235.0n ± ∞ ¹  +899.06% (p=0.008 n=5)

It doesn't show up most of the time when scraping, as the initial construction of labels is cached.  But when reading the WAL or creating a large set of new series, this overhead will be painful.
Also when receiving data via remote-write or OTLP, every series is constructed fresh - those users will pay the overhead continuously.

So, I don't think we could have this on by default.

Do you find that this version is better than `-tags dedupelabels` ?  On what dimension(s) ?
Maybe the idea works better as a modification to that, rather than a modification to `stringlabels` ?