@bboreham 2025-04-22
Hello from the bug-scrub!

I'm personally still unable to accept a benchmark result going 70% slower.  In our cloud bills CPU is about 4x the price of RAM.

Discussed at the bug-scrub: we could add this variant under another build tag, letting people in similar situations to yourself get the benefit.