@bboreham 2025-05-02
Still thinking whether doesn't work better as a mod to `dedupelabels`.  Basically you have one symbol-table which you never change, plus a fall-back to store strings that aren't in the table.